Migrate tests to the right directory
